pyproject.toml
[project]
name = "consolidate_markdown"
version = "1.0.0"
description = "Consolidate Markdown files from multiple sources with AI-powered image analysis"
requires-python = ">=3.12"
dependencies = [
  "tomli>=2.0.1",
  "openai>=1.0.0",
  "pathlib>=1.0.1",
  "pillow>=10.0.0",
  "cairosvg>=2.7.0",
  "pandas>=2.1.0",
  "tabulate>=0.9.0",
  "openpyxl>=3.1.0",
  "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=23.0.0", "isort>=5.0.0", "mypy>=1.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
